test: split slow Go test packages
Large internal/server and internal/workspace test packages serialized several minutes of work behind one package process. This limited Go's package-level scheduling even when other cores were idle.

Keep self-contained request and Git status tests in independently scheduled packages. The test inventory stays intact while the measured package critical path drops from 258 seconds to 177 seconds with four test slots.
