+++ internal/workspace/gitstatustest/branch_sync_test.go
@@ -1,4 +1,4 @@
-package workspace
+package gitstatustest
 
 import (
 	"context"
