+++ internal/workspace/gitstatustest/divergence_test.go
@@ -1,4 +1,4 @@
-package workspace
+package gitstatustest
 
 import (
 	"os"
